refactor(Promotion) : move Promotion class
- create store.product.promotion package to manage promotion only
